+++ packages/stream_chat_flutter/test/src/theme/gallery_footer_theme_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' hide TextTheme;
 import 'package:flutter_test/flutter_test.dart';
-import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 class MockStreamChatClient extends Mock implements StreamChatClient {}