+++ packages/stream_chat/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Upcoming
 
+✅ Added
+
+- Added `Filter.notExists`.
+
+🐞 Fixed
+
+- [[#710]](https://github.com/GetStream/stream-chat-flutter/issues/710) Fixed JWT requiring using `String` as id.
+
+## 3.0.0
+
 🛑️ Breaking Changes from `2.2.1`
 
 - Added 6 new methods in `ChatPersistenceClient`.
@@ -16,13 +26,16 @@
 - Added support for `next`, `previous` value pagination in `client.search`
   , [read more.](https://getstream.io/chat/docs/other-rest/search/#pagination)
 - `Attachment` class now has a `fileSize` and `mimeType` property. Setting a `file` will also set the `file_size`
-  , `mime_type` key on `extraData`, so `attachment.fileSize`, `attachment.mimetype` and `attachment.extraData['file_size']`
+  , `mime_type` key on `extraData`, so `attachment.fileSize`, `attachment.mimetype`
+  and `attachment.extraData['file_size']`
   , `attachment.extraData['mime_type]` is same respectively.
 
 🐞 Fixed
 
 - [[#659]](https://github.com/GetStream/stream-chat-flutter/issues/659) Fixed unread count not updating correctly.
 - Fix `Filter.empty()` json encoding.
+- [[#700]](https://github.com/GetStream/stream-chat-flutter/issues/700) Connecting user without providing `name`
+  uses `id` instead for setting `user.name`.
 
 ## 2.2.1